# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "contact": {
      "name": "NIEMOpen",
      "url": "https://niemopen.org"
    },
    "description": "Community support for common NIEM tool functionality to browse and search the model, build and migrate subsets, and check conformance",
    "title": "NIEM API 2.0",
    "version": "2.0-alpha08.1"
  },
  "servers": [
    {
      "url": "https://api.niemopen.org/v2",
      "description": "Production server"
    },
    {
      "url": "https://api.niemopen.org/mep",
      "description": "MEP instance server"
    },
    {
      "url": "http://localhost:8080/v2",
      "description": "Local machine"
    }
  ],
...
  "paths": {
...
    "/stewards/{stewardKey}/models/{modelKey}/versions/current": {
      "get": {
        "description": "Get the current instance of the version with the given fields.",
        "operationId": "getCurrent_4",
        "parameters": [
          {
            "in": "path",
            "name": "stewardKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "modelKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Version"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get the current instance of the version with the given fields.",
        "tags": [
          "Data-3: Versions"
        ]
      }
    },
    "/stewards/{stewardKey}/models/{modelKey}/versions/{versionNumber}": {
      "get": {
        "description": "Gets the version with the given fields.",
        "operationId": "getVersion",
        "parameters": [
          {
            "in": "path",
            "name": "stewardKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "modelKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "versionNumber",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Version"
                }
              }
            },
            "description": "OK"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Gets the version with the given fields.",
        "tags": [
          "Data-3: Versions"
        ]
      }
    },
...
    "/stewards/{stewardKey}/models/{modelKey}/versions/{versionNumber}/history": {
      "get": {
        "description": "Get the history of the version with the given fields.",
        "operationId": "getHistory_5",
        "parameters": [
          {
            "in": "path",
            "name": "stewardKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "modelKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "versionNumber",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "- True to return a result from a pre-release if applicable;\n     false to iterate until an official version is reached.",
            "in": "query",
            "name": "includePreRelease",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Version"
                  }
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get the history of the version with the given fields.",
        "tags": [
          "Data-3: Versions"
        ]
      }
    },
    "/stewards/{stewardKey}/models/{modelKey}/versions/{versionNumber}/namespaces": {
      "get": {
        "description": "Get all namespaces from a version of a model.",
        "operationId": "getVersionNamespaces",
        "parameters": [
          {
            "in": "path",
            "name": "stewardKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "modelKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "versionNumber",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Namespace"
                  }
                }
              }
            },
            "description": "OK"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get all namespaces from a version of a model.",
        "tags": [
          "Data-4: Namespace"
        ]
      }
    },
    "/stewards/{stewardKey}/models/{modelKey}/versions/{versionNumber}/namespaces/count": {
      "get": {
        "description": "Count all namespaces in the version with the given fields.",
        "operationId": "countNamespaces",
        "parameters": [
          {
            "in": "path",
            "name": "stewardKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "modelKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "versionNumber",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "core",
                "domain",
                "code",
                "adapter",
                "auxiliary",
                "external",
                "utility",
                "core_supplement",
                "domain_update",
                "extension",
                "exchange",
                "built_in",
                "other"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Count all namespaces in the version with the given fields.",
        "tags": [
          "Data-4: Namespace"
        ]
      }
    },
    "/stewards/{stewardKey}/models/{modelKey}/versions/{versionNumber}/namespaces/formats/cmf": {
      "get": {
        "description": "Get all namespaces as CMF from a version of a model.",
        "operationId": "getVersionNamespacesCmf",
        "parameters": [
          {
            "in": "path",
            "name": "stewardKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "modelKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "versionNumber",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "mediaType",
            "required": false,
            "schema": {
              "type": "string",
              "default": "json",
              "enum": [
                "json",
                "xml"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get all namespaces as CMF from a version of a model.",
        "tags": [
          "Data-4: Namespace"
        ]
      }
    },
    "/stewards/{stewardKey}/models/{modelKey}/versions/{versionNumber}/namespaces/{prefix}": {
      "get": {
        "description": "Get a namespace with the given fields.",
        "operationId": "getNamespace",
        "parameters": [
          {
            "in": "path",
            "name": "stewardKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "modelKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "versionNumber",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "prefix",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Namespace"
                }
              }
            },
            "description": "OK"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get a namespace with the given fields.",
        "tags": [
          "Data-4: Namespace"
        ]
      }
    },
    "/stewards/{stewardKey}/models/{modelKey}/versions/{versionNumber}/namespaces/{prefix}/current": {
      "get": {
        "description": "Get the current version of the namespace with the given fields.",
        "operationId": "getNext_6",
        "parameters": [
          {
            "in": "path",
            "name": "stewardKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "modelKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "versionNumber",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "prefix",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Namespace"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get the current version of the namespace with the given fields.",
        "tags": [
          "Data-4: Namespace"
        ]
      }
    },
...
    "/stewards/{stewardKey}/models/{modelKey}/versions/{versionNumber}/namespaces/{prefix}/history": {
      "get": {
        "description": "Get the history of the namespace with the given fields.",
        "operationId": "getHistory_4",
        "parameters": [
          {
            "in": "path",
            "name": "stewardKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "modelKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "versionNumber",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "prefix",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "- True to return a result from a pre-release if applicable;\n     false to iterate until an official version is reached.",
            "in": "query",
            "name": "includePreRelease",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Namespace"
                  }
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get the history of the namespace with the given fields.",
        "tags": [
          "Data-4: Namespace"
        ]
      }
    },
    "/stewards/{stewardKey}/models/{modelKey}/versions/{versionNumber}/namespaces/{prefix}/next": {
      "get": {
        "description": "Get the next version of the namespace with the given fields.",
        "operationId": "getNext_5",
        "parameters": [
          {
            "in": "path",
            "name": "stewardKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "modelKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "versionNumber",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "prefix",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "- True to return a result from a pre-release if applicable;\n     false to iterate until an official version is reached.",
            "in": "query",
            "name": "includePreRelease",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Namespace"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get the next version of the namespace with the given fields.",
        "tags": [
          "Data-4: Namespace"
        ]
      }
    },
    "/stewards/{stewardKey}/models/{modelKey}/versions/{versionNumber}/namespaces/{prefix}/prev": {
      "get": {
        "description": "Get the previous version of the namespace with the given fields.",
        "operationId": "getPrev_5",
        "parameters": [
          {
            "in": "path",
            "name": "stewardKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "modelKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "versionNumber",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "prefix",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "- True to return a result from a pre-release if applicable;\n     false to iterate until an official version is reached.",
            "in": "query",
            "name": "includePreRelease",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Namespace"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get the previous version of the namespace with the given fields.",
        "tags": [
          "Data-4: Namespace"
        ]
      }
    },
...
    "/stewards/{stewardKey}/models/{modelKey}/versions/{versionNumber}/next": {
      "get": {
        "description": "Get the next version of the version with the given fields.",
        "operationId": "getNext_4",
        "parameters": [
          {
            "in": "path",
            "name": "stewardKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "modelKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "versionNumber",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "- True to return a result from a pre-release if applicable;\n     false to iterate until an official version is reached.",
            "in": "query",
            "name": "includePreRelease",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Version"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get the next version of the version with the given fields.",
        "tags": [
          "Data-3: Versions"
        ]
      }
    },
    "/stewards/{stewardKey}/models/{modelKey}/versions/{versionNumber}/prev": {
      "get": {
        "description": "Get the previous version of the version with the given fields.",
        "operationId": "getPrev_4",
        "parameters": [
          {
            "in": "path",
            "name": "stewardKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "modelKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "versionNumber",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "- True to return a result from a pre-release if applicable;\n     false to iterate until an official version is reached.",
            "in": "query",
            "name": "includePreRelease",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Version"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get the previous version of the version with the given fields.",
        "tags": [
          "Data-3: Versions"
        ]
      }
    },
...
    "/stewards/{stewardKey}/models/{modelKey}/versions/{versionNumber}/properties/{qname}/current": {
      "get": {
        "description": "Get the current version of the property with the given fields.",
        "operationId": "getCurrent_3",
        "parameters": [
          {
            "in": "path",
            "name": "stewardKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "modelKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "versionNumber",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "qname",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Property"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get the current version of the property with the given fields.",
        "tags": [
          "Data-5: Properties"
        ]
      }
    },
...
    "/stewards/{stewardKey}/models/{modelKey}/versions/{versionNumber}/properties/{qname}/history": {
      "get": {
        "description": "Get the history of the property with the given fields.",
        "operationId": "getHistory_3",
        "parameters": [
          {
            "in": "path",
            "name": "stewardKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "modelKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "versionNumber",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "qname",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "- True to return a result from a pre-release if applicable;\n     false to iterate until an official version is reached.",
            "in": "query",
            "name": "includePreRelease",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Property"
                  }
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get the history of the property with the given fields.",
        "tags": [
          "Data-5: Properties"
        ]
      }
    },
...
    "/stewards/{stewardKey}/models/{modelKey}/versions/{versionNumber}/types/{qname}/current": {
      "get": {
        "description": "Get the current version of the type with the given fields.",
        "operationId": "getCurrent_2",
        "parameters": [
          {
            "in": "path",
            "name": "stewardKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "modelKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "versionNumber",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "qname",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Type"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get the current version of the type with the given fields.",
        "tags": [
          "Data-6: Types"
        ]
      }
    },
...
  },
...
}