# src/main/java/gov/niem/tools/api/db/namespace/NamespaceController.java
package gov.niem.tools.api.db.namespace;

import gov.niem.tools.api.core.config.Config.AppMediaType;
import gov.niem.tools.api.core.utils.CmfUtils;
import gov.niem.tools.api.db.ServiceHub;
import gov.niem.tools.api.db.base.AddModelReason;
import gov.niem.tools.api.db.exceptions.EntityNotFoundException;

import org.mitre.niem.cmf.Model;

import io.swagger.v3.oas.annotations.Hidden;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import java.util.List;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.ResponseStatus;
import org.springframework.web.bind.annotation.RestController;

/**
 * REST controller for namespaces.
 */
@Validated
@RestController
@RequestMapping("/stewards/{stewardKey}/models/{modelKey}/versions/{versionNumber}")
@Tag(
    name = "Data-4: Namespace",
    description = "A collection of properties and types managed by an authoritative source.")
public class NamespaceController {

  @Autowired
  ServiceHub hub;

  /**
   * Get a namespace with the given fields.
   *
   * @example http://api.niemopen.org/v2/stewards/niem/models/model/versions/5.2/namespaces/nc
   */
  @GetMapping("/namespaces/{prefix}")
  @ResponseStatus(code = HttpStatus.OK)
  @ApiResponse(responseCode = "422", description = "Unprocessable Entity", content = @Content)
  public Namespace getNamespace(
      @PathVariable String stewardKey,
      @PathVariable String modelKey,
      @PathVariable String versionNumber,
      @PathVariable String prefix) throws Exception {
    return hub.namespaces.findOne(stewardKey, modelKey, versionNumber, prefix);
  }

  /**
   * Get a namespace in CMF with the given fields.
   */
  @Hidden
  @GetMapping("/namespaces/{prefix}/formats/cmf")
  @ResponseStatus(code = HttpStatus.OK)
  @ApiResponse(responseCode = "422", description = "Unprocessable Entity", content = @Content)
  public Object getNamespaceCmf(
      @PathVariable String stewardKey,
      @PathVariable String modelKey,
      @PathVariable String versionNumber,
      @PathVariable String prefix,
      @RequestParam(required = false, defaultValue = "json") AppMediaType mediaType)
      throws Exception {
    Namespace namespace = hub.namespaces.findOne(stewardKey, modelKey, versionNumber, prefix);
    org.mitre.niem.cmf.Model cmfModel = new Model();
    namespace.addToCmfModel(cmfModel, false, AddModelReason.REPRESENTATION, null);
    // TODO: Add properties, types, and dependencies to namespace CMF
    return CmfUtils.generateString(cmfModel, mediaType);
  }

  /**
   * Get all namespaces from a version of a model.
   *
   * @example http://api.niemopen.org/v2/stewards/niem/models/model/versions/5.2/namespaces
   */
  @GetMapping("/namespaces")
  @ResponseStatus(code = HttpStatus.OK)
  @ApiResponse(responseCode = "422", description = "Unprocessable Entity", content = @Content)
  public List<Namespace> getVersionNamespaces(
      @PathVariable String stewardKey,
      @PathVariable String modelKey,
      @PathVariable String versionNumber) throws Exception {
    return hub.namespaces.findByVersion(stewardKey, modelKey, versionNumber);
  }

  /**
   * Get all namespaces as CMF from a version of a model.
   */
  @GetMapping("/namespaces/formats/cmf")
  @ResponseStatus(code = HttpStatus.OK)
  @ApiResponse(responseCode = "422", description = "Unprocessable Entity", content = @Content)
  public Object getVersionNamespacesCmf(
      @PathVariable String stewardKey,
      @PathVariable String modelKey,
      @PathVariable String versionNumber,
      @RequestParam(required = false, defaultValue = "json") AppMediaType mediaType)
      throws Exception {

    // Get results
    List<Namespace> namespaces = hub.namespaces.findByVersion(stewardKey, modelKey, versionNumber);

    // Convert results to CMF
    org.mitre.niem.cmf.Model cmfModel = new Model();
    for (Namespace namespace : namespaces) {
      namespace.addToCmfModel(cmfModel, false, AddModelReason.REPRESENTATION, null);
    }
    return CmfUtils.generateString(cmfModel, mediaType);
  }

  /**
   * Count all namespaces in the version with the given fields.
   */
  @GetMapping("/namespaces/count")
  @ResponseStatus(code = HttpStatus.OK)
  public long countNamespaces(
      @PathVariable String stewardKey,
      @PathVariable String modelKey,
      @PathVariable String versionNumber,
      @RequestParam(required = false) Namespace.Category category) throws EntityNotFoundException {
    return hub.namespaces.count(stewardKey, modelKey, versionNumber, category);
  }

  /**
   * Get the previous version of the namespace with the given fields.
   *
   * @param includePreRelease - True to return a result from a pre-release if applicable;
   *     false to iterate until an official version is reached.
   */
  @GetMapping("/namespaces/{prefix}/prev")
  @ResponseStatus(code = HttpStatus.OK)
  public Namespace getPrev(
      @PathVariable String stewardKey,
      @PathVariable String modelKey,
      @PathVariable String versionNumber,
      @PathVariable String prefix,
      @RequestParam(required = false, defaultValue = "false") boolean includePreRelease)
      throws Exception {
    Namespace namespace = hub.namespaces.findOne(stewardKey, modelKey, versionNumber, prefix);
    return hub.namespaces.getPrev(namespace, includePreRelease);
  }

  /**
   * Get the next version of the namespace with the given fields.
   *
   * @param includePreRelease - True to return a result from a pre-release if applicable;
   *     false to iterate until an official version is reached.
   */
  @GetMapping("/namespaces/{prefix}/next")
  @ResponseStatus(code = HttpStatus.OK)
  public Namespace getNext(
      @PathVariable String stewardKey,
      @PathVariable String modelKey,
      @PathVariable String versionNumber,
      @PathVariable String prefix,
      @RequestParam(required = false, defaultValue = "false") boolean includePreRelease)
      throws Exception {
    Namespace namespace = hub.namespaces.findOne(stewardKey, modelKey, versionNumber, prefix);
    return hub.namespaces.getNext(namespace, includePreRelease);
  }

  /**
   * Get the current version of the namespace with the given fields.
   */
  @GetMapping("/namespaces/{prefix}/current")
  @ResponseStatus(code = HttpStatus.OK)
  public Namespace getNext(
      @PathVariable String stewardKey,
      @PathVariable String modelKey,
      @PathVariable String versionNumber,
      @PathVariable String prefix) throws Exception {
    Namespace namespace = hub.namespaces.findOne(stewardKey, modelKey, versionNumber, prefix);
    return namespace.getCurrent();
  }

  /**
   * Get the history of the namespace with the given fields.
   *
   * @param includePreRelease - True to return a result from a pre-release if applicable;
   *     false to iterate until an official version is reached.
   */
  @GetMapping("/namespaces/{prefix}/history")
  @ResponseStatus(code = HttpStatus.OK)
  public List<Namespace> getHistory(
      @PathVariable String stewardKey,
      @PathVariable String modelKey,
      @PathVariable String versionNumber,
      @PathVariable String prefix,
      @RequestParam(required = false, defaultValue = "false") boolean includePreRelease)
      throws Exception {
    Namespace namespace = hub.namespaces.findOne(stewardKey, modelKey, versionNumber, prefix);
    return hub.namespaces.getHistory(namespace, includePreRelease);
  }

  // @PostMapping("/namespaces")
  // public ResponseEntity<String> postNamespace(@PathVariable String stewardKey,
  //     @PathVariable String modelKey, @PathVariable String versionNumber, Namespace namespace)
  //     throws Exception {
  //   String message = hub.namespaces.add(stewardKey, modelKey, versionNumber, namespace);
  //   return AppUtils.getResponseOkString(message);
  // }

}